# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/paper/{id}": {
      "get": {
        "operationId": "PaperController_getById",
        "summary": "",
        "description": "Get paper by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      },
      "patch": {
        "operationId": "PaperController_update",
        "summary": "",
        "description": "Update paper",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      },
      "delete": {
        "operationId": "PaperController_delete",
        "summary": "",
        "description": "Delete paper",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      }
    },
    "/paper": {
      "get": {
        "operationId": "PaperController_getAll",
        "summary": "",
        "description": "Get all papers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      },
      "post": {
        "operationId": "PaperController_create",
        "summary": "",
        "description": "Create new paper",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      }
    },
    "/paper/search": {
      "get": {
        "operationId": "PaperController_getByFilter",
        "summary": "",
        "description": "Get papers with filter",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      }
    },
    "/paper/my/{id}": {
      "get": {
        "operationId": "PaperController_getMy",
        "summary": "",
        "description": "Get all users papers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDTO"
                }
              }
            }
          }
        },
        "tags": [
          "paper"
        ]
      }
    },
...
    "/paper/{filePath}": {
      "get": {
        "operationId": "PaperController_seeUploadedFile",
        "summary": "",
        "description": "Download paper",
        "parameters": [
          {
            "name": "filePath",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paper"
        ]
      }
    },
    "/review/{id}": {
      "get": {
        "operationId": "ReviewController_getById",
        "summary": "",
        "description": "Get review by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        },
        "tags": [
          "review"
        ]
      },
      "patch": {
        "operationId": "ReviewController_update",
        "summary": "",
        "description": "Update review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        },
        "tags": [
          "review"
        ]
      },
      "delete": {
        "operationId": "ReviewController_delete",
        "summary": "",
        "description": "Delete review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        },
        "tags": [
          "review"
        ]
      }
    },
    "/review": {
      "get": {
        "operationId": "ReviewController_getAll",
        "summary": "",
        "description": "Get all reviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        },
        "tags": [
          "review"
        ]
      },
      "post": {
        "operationId": "ReviewController_create",
        "summary": "",
        "description": "Create new review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        },
        "tags": [
          "review"
        ]
      }
    },
    "/reputation-model/{id}": {
      "get": {
        "operationId": "ReputationModelController_getById",
        "summary": "",
        "description": "Get rm by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationModelDTO"
                }
              }
            }
          }
        },
        "tags": [
          "reputation-model"
        ]
      },
      "patch": {
        "operationId": "ReputationModelController_update",
        "summary": "",
        "description": "Update rm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReputationModelUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationModelDTO"
                }
              }
            }
          }
        },
        "tags": [
          "reputation-model"
        ]
      },
      "delete": {
        "operationId": "ReputationModelController_delete",
        "summary": "",
        "description": "Delete rm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationModelDTO"
                }
              }
            }
          }
        },
        "tags": [
          "reputation-model"
        ]
      }
    },
    "/reputation-model/{id}/scorings": {
      "get": {
        "operationId": "ReputationModelController_getModelScorings",
        "summary": "",
        "description": "Get all scorings of a reputation model",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringDTO"
                }
              }
            }
          }
        },
        "tags": [
          "reputation-model"
        ]
      }
    },
    "/reputation-model": {
      "get": {
        "operationId": "ReputationModelController_getAll",
        "summary": "",
        "description": "Get all rms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationModelDTO"
                }
              }
            }
          }
        },
        "tags": [
          "reputation-model"
        ]
      },
      "post": {
        "operationId": "ReputationModelController_create",
        "summary": "",
        "description": "Create new rm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReputationModelCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationModelDTO"
                }
              }
            }
          }
        },
        "tags": [
          "reputation-model"
        ]
      }
    },
...
  },
  "info": {
    "title": "BlockTrust API",
    "description": "The BlockTrust API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "HistoryLogLine": {
        "type": "object",
        "properties": {}
      },
      "Author": {
        "type": "object",
        "properties": {}
      },
      "FileMetadata": {
        "type": "object",
        "properties": {}
      },
      "PaperDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "abstract": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "submissionDate": {
            "format": "date-time",
            "type": "string"
          },
          "publicationDate": {
            "format": "date-time",
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transactions": {
            "type": "object"
          },
          "doi": {
            "type": "string"
          },
          "disciplines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "references": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "number"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "socialComments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canBeReviewed": {
            "type": "boolean"
          },
          "gitRepository": {
            "type": "string"
          },
          "openAireIdentifier": {
            "type": "string"
          },
          "views": {
            "type": "number"
          },
          "isLatestVersion": {
            "type": "boolean"
          },
          "paperType": {
            "enum": [
              "book",
              "book section",
              "conference paper",
              "article",
              "patent",
              "preprint",
              "report",
              "dataset",
              "software documentation",
              "thesis",
              "technical note",
              "working paper",
              "policy report",
              "registered report",
              "proposal",
              "review article",
              "video",
              "other"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "pending approval",
              "published",
              "preprint",
              "rejected",
              "merged"
            ],
            "type": "string"
          },
          "reviewType": {
            "type": "string",
            "enum": [
              "open review"
            ]
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryLogLine"
            }
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            }
          },
          "publicationFile": {
            "$ref": "#/components/schemas/FileMetadata"
          }
        },
        "required": [
          "_id",
          "creator",
          "title",
          "abstract",
          "createdAt",
          "updatedAt",
          "submissionDate",
          "publicationDate",
          "keywords",
          "transactions",
          "doi",
          "disciplines",
          "references",
          "version",
          "images",
          "socialComments",
          "canBeReviewed",
          "gitRepository",
          "openAireIdentifier",
          "views",
          "isLatestVersion",
          "paperType",
          "status",
          "reviewType",
          "history",
          "authors",
          "publicationFile"
        ]
      },
      "PaperCreateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "creator"
        ]
      },
      "PaperUpdateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "abstract": {
            "type": "string"
          },
          "submissionDate": {
            "format": "date-time",
            "type": "string"
          },
          "publicationDate": {
            "format": "date-time",
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transactions": {
            "type": "object"
          },
          "doi": {
            "type": "string"
          },
          "disciplines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "references": {
            "type": "object"
          },
          "canBeReviewed": {
            "type": "boolean"
          },
          "gitRepository": {
            "type": "string"
          },
          "paperType": {
            "type": "string",
            "enum": [
              "book",
              "book section",
              "conference paper",
              "article",
              "patent",
              "preprint",
              "report",
              "dataset",
              "software documentation",
              "thesis",
              "technical note",
              "working paper",
              "policy report",
              "registered report",
              "proposal",
              "review article",
              "video",
              "other"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "pending approval",
              "published",
              "preprint",
              "rejected",
              "merged"
            ]
          },
          "reviewType": {
            "type": "string",
            "enum": [
              "open review"
            ]
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            }
          }
        }
      },
      "ReviewDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "publicationDate": {
            "format": "date-time",
            "type": "string"
          },
          "wasInvited": {
            "type": "boolean"
          },
          "transactions": {
            "type": "object"
          },
          "doi": {
            "type": "string"
          },
          "socialComments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canBeReviewed": {
            "type": "boolean"
          },
          "reward": {
            "type": "number"
          },
          "revealReviewerIdentity": {
            "type": "boolean"
          },
          "paper": {
            "type": "string"
          },
          "decision": {
            "enum": [
              "accepted",
              "minor revision",
              "major revision"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "published",
              "pending approval"
            ],
            "type": "string"
          },
          "file": {
            "$ref": "#/components/schemas/FileMetadata"
          },
          "kind": {
            "enum": [
              "peer review",
              "copy editing"
            ],
            "type": "string"
          }
        },
        "required": [
          "_id",
          "creator",
          "title",
          "createdAt",
          "updatedAt",
          "publicationDate",
          "socialComments",
          "canBeReviewed",
          "revealReviewerIdentity",
          "paper",
          "status",
          "kind"
        ]
      },
      "ReviewCreateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "paper": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "creator",
          "paper"
        ]
      },
      "ReviewUpdateDTO": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "pending approval"
            ]
          },
          "reward": {
            "type": "number"
          },
          "revealReviewerIdentity": {
            "type": "boolean"
          },
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "minor revision",
              "major revision"
            ]
          },
          "doi": {
            "type": "string"
          },
          "transactions": {
            "type": "object"
          },
          "publicationDate": {
            "format": "date-time",
            "type": "string"
          },
          "wasInvited": {
            "type": "boolean"
          }
        }
      },
      "ReputationModelDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "sourceCodeURL": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "_id",
          "creator",
          "name",
          "description",
          "version",
          "sourceCodeURL",
          "provider",
          "createdAt",
          "updatedAt"
        ]
      },
      "ScoringDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "reputationModel": {
            "type": "string"
          },
          "paper": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "_id",
          "reputationModel",
          "paper",
          "score",
          "timestamp"
        ]
      },
      "ReputationModelCreateDTO": {
        "type": "object",
        "properties": {
          "creator": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sourceCodeURL": {
            "type": "string"
          }
        },
        "required": [
          "creator",
          "name",
          "sourceCodeURL"
        ]
      },
      "ReputationModelUpdateDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "sourceCodeURL": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "version",
          "sourceCodeURL",
          "provider"
        ]
      },
...
    }
  }
}